requirements.txt
python-dotenv 
ndjson
influxdb
paho-mqtt
dotenv